requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==22.2.0
blinker==1.9.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.8
crashtest==0.4.1
cryptography==39.0.2
Deprecated==1.2.18
distlib==0.3.6
distro==1.9.0
dulwich==0.21.3
filelock==3.9.0
Flask==3.1.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
itsdangerous==2.2.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.5
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==3.0.2
more-itertools==9.1.0
msgpack==1.0.5
netaddr==1.3.0
packaging==23.0
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
py3xui==0.3.4
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.1
python-telegram-bot==21.10
rapidfuzz==2.15.1
redis==5.2.1
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
sniffio==1.3.1
tomlkit==0.11.6
trove-classifiers==2023.3.9
typing_extensions==4.12.2
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
waitress==3.0.2
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
yookassa==3.5.0
zipp==3.15.0
email_validator~=2.2.0